fix(pie): fix some comments
