+++ src/components/pie-chart/pie-chart.ts
@@ -3,9 +3,10 @@ import { css, html, LitElement, TemplateResult } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { merge } from 'lodash-es';
 import { COMPONENT_PREFIX, ThemeKey, themes } from '../../core';
+import { chartOptions } from '../../core/chart-options';
 import { ChartA11y, ChartLegendA11y, legendClickHandler } from '../../core/plugins';
 import { LegendClickData } from '../../core/plugins/plugin.types';
-import { chartOptions } from '../../types';
+import { getCurrentTheme } from '../../core/utils';
 import { defaultPieChartOptions } from './pie-chart.options';
 import { CenterValue } from './pie-chart.plugins';
 import { PieChartOptions } from './pie-chart.types';
@@ -156,7 +157,7 @@ class PieChart extends LitElement {
   }
 
   private getBackgroundColor(chartOptions: PieChartOptions): string[] | undefined {
-    return typeof chartOptions?.theme === 'string' ? themes.get(chartOptions?.theme as keyof typeof ThemeKey) : chartOptions?.theme;
+    return typeof chartOptions?.theme === 'string' ? themes.get(chartOptions?.theme as keyof typeof ThemeKey) : getCurrentTheme().colors;
   }
 
   updateChart(): void {
