+++ src/core/chart-options.ts
@@ -1,0 +1,13 @@
+import { ChartOptions } from '../types';
+
+const chartOptions: ChartOptions = {
+  responsive: true,
+  chartLabel: '',
+  fontColor: '#000',
+  fontFamily: 'Helvetica',
+  aspectRatio: 1.6,
+  legendLabelsHeight: 12,
+  legendLabelsWidth: 12,
+};
+
+export { chartOptions };
