+++ src/types/chart.types.ts
@@ -1,5 +1,3 @@
-import { defaultTheme } from '../core';
-
 export type LayoutPosition = 'left' | 'top' | 'right' | 'bottom' | 'center' | 'chartArea' | { [scaleId: string]: number };
 
 export interface ChartOptions {
@@ -19,16 +17,3 @@ export interface ChartOptions {
   paddingTop?: number;
   paddingBottom?: number;
 }
-
-const chartOptions: ChartOptions = {
-  responsive: true,
-  chartLabel: '',
-  theme: defaultTheme,
-  fontColor: '#000',
-  fontFamily: 'Helvetica',
-  aspectRatio: 1.6,
-  legendLabelsHeight: 12,
-  legendLabelsWidth: 12,
-};
-
-export { chartOptions };
